extract reporter to another file
