+++ lib/jasminerice-runner.rb
@@ -1,3 +1,5 @@
+require 'jasminerice-runner/reporter'
+
 module Jasminerice
   class Runner
     cattr_accessor :capybara_driver
